Download osm data from geofabric

Download osm data from geofabric

get_geofabric(name = "west-yorkshire", layer = "lines",
  attributes = make_additional_attributes(layer = layer),
  download_directory = tempdir(), ask = TRUE, max_dist = 3)

Arguments

name
Name of the geofabric zone to download

layer
Character string telling `sf` which OSM layer to import. One of `points`, `lines` (the default), `multilinestrings`, `multipolygons` or `other_relations`

attributes
Vector of character strings naming attributes to import

download_directory
Where to download the data? `tempdir()` by default.

ask
Should the user be asked before downloading the file?

max_dist
What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download?

Examples

get_geofabric("isle of man")

#> No exact matching geofabric zone. Best match is Isle of Man (2.2 MB)

#> Downloading http://download.geofabrik.de/europe/isle-of-man-latest.osm.pbf to 
#> /tmp/RtmpKQv95T/isle of man.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpKQv95T/ini_new.ini)

# \donttest{
get_geofabric("andorra")

#> No exact matching geofabric zone. Best match is Andorra (1.5 MB)

#> Downloading http://download.geofabrik.de/europe/andorra-latest.osm.pbf to 
#> /tmp/RtmpKQv95T/andorra.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpKQv95T/ini_new.ini)

get_geofabric("west-yorkshire")

#> No exact matching geofabric zone. Best match is West Yorkshire (28.7 MB)

#> Downloading http://download.geofabrik.de/europe/great-britain/england/west-yorkshire-latest.osm.pbf to 
#> /tmp/RtmpKQv95T/west-yorkshire.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpKQv95T/ini_new.ini)

# user asked to choose closest match when interactive
get_geofabric("kdljfdl")

#> No exact matching geofabric zone. Best match is Poland (1.1 GB)

#> Downloading http://download.geofabrik.de/europe/poland-latest.osm.pbf to 
#> /tmp/RtmpKQv95T/kdljfdl.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpKQv95T/ini_new.ini)

#> Warning: GDAL Error 1: Too many features have accumulated in points layer. Use OGR_INTERLEAVED_READING=YES mode

# get zone associated with a point
name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
get_geofabric(name)

#> although coordinates are longitude/latitude, st_within assumes that they are planar

#> although coordinates are longitude/latitude, st_contains assumes that they are planar

#> The place is within these geofabrik zones: Europe, Great Britain, Britain and Ireland, England, Lincolnshire
#> Selecting the smallest: Lincolnshire

#> Downloading http://download.geofabrik.de/europe/great-britain/england/lincolnshire-latest.osm.pbf to 
#> /tmp/RtmpKQv95T/c(0, 53).osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpKQv95T/ini_new.ini)

# }