Read pbf files with additional attributes

Read pbf files with additional attributes

read_pbf(dsn, layer = "lines",
  attributes = make_additional_attributes(layer = layer),
  ini_file = NULL, append = TRUE)

Arguments

dsn
The location of the file

layer
Which layer to read in? One of "points" "lines" "multipolygons" "multilinestrings" or "other_relations"

attributes
Vector of character strings naming attributes to import

ini_file
A modified version of https://github.com/OSGeo/gdal/raw/master/gdal/data/osmconf.ini

append
Should the columns named in `attributes` be appended to the default columns? `TRUE` by default.

Examples

# \donttest{
pbf_url = geofabric_zones$pbf_url[geofabric_zones$name == "Isle of Wight"]
f = file.path(tempdir(), "test.osm.pbf")
download.file(pbf_url, f)
res = read_pbf(f)

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/Rtmp6f1yaK/ini_new.ini)

names(res)

#>  [1] "osm_id"     "name"       "highway"    "waterway"   "aerialway" 
#>  [6] "barrier"    "man_made"   "maxspeed"   "oneway"     "building"  
#> [11] "surface"    "landuse"    "natural"    "start_date" "wall"      
#> [16] "service"    "lanes"      "layer"      "tracktype"  "bridge"    
#> [21] "foot"       "bicycle"    "lit"        "railway"    "footway"   
#> [26] "z_order"    "other_tags" "geometry"  

res = read_pbf(f, layer = "points")

#> Old attributes: attributes=name,barrier,highway,ref,address,is_in,place,man_made

#> New attributes: attributes=name,barrier,highway,ref,address,is_in,place,man_made,building,natural,surface,source,power,amenity,shop,operator

#> Using ini file that can can be edited with file.edit(/tmp/Rtmp6f1yaK/ini_new.ini)

names(res)

#>  [1] "osm_id"     "name"       "barrier"    "highway"    "ref"       
#>  [6] "address"    "is_in"      "place"      "man_made"   "building"  
#> [11] "natural"    "surface"    "source"     "power"      "amenity"   
#> [16] "shop"       "operator"   "other_tags" "geometry"  

# }