Download osm data from geofabric

Download osm data from geofabric

get_geofabric(name = "west-yorkshire", layer = "lines",
  attributes = make_additional_attributes(layer = layer),
  download_directory = gf_download_directory(), ask = TRUE,
  max_dist = 5, op = sf::st_contains)

Arguments

name
Name or spatial object of the geofabric zone to download

layer
Character string telling `sf` which OSM layer to import. One of `points`, `lines` (the default), `multilinestrings`, `multipolygons` or `other_relations`

attributes
Vector of character strings naming attributes to import

download_directory
Where to download the data? `tempdir()` by default. If you want to download your data into a persistent directory, set `GF_DOWNLOAD_DIRECTORY=/path/for/osm/data` in your `.Renviron` file, e.g. with `usethis::edit_r_environ()`.

ask
Should the user be asked before downloading the file?

max_dist
What is the maximum distance in fuzzy matching to tolerate before asking the user to select which zone to download?

op
The binary spatial predicate used to identify the smallest geofabric zones that matches the simple feature input in `name`

Examples

get_geofabric("isle of man")

#> No exact matching geofabric zone. Best match is Isle of Man (2.2 MB)

#> Downloading http://download.geofabrik.de/europe/isle-of-man-latest.osm.pbf to 
#> /tmp/RtmpWjPinl/Isle of Man.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpWjPinl/ini_new.ini)

# \donttest{
get_geofabric(name = "andorra") # try other names, e.g. name = "west-yorkshire"

#> No exact matching geofabric zone. Best match is Andorra (1.5 MB)

#> Downloading http://download.geofabrik.de/europe/andorra-latest.osm.pbf to 
#> /tmp/RtmpWjPinl/Andorra.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpWjPinl/ini_new.ini)

# user asked to choose closest match when interactive
get_geofabric("kdljfdl", ask = FALSE)

#> No exact matching geofabric zone. Best match is Poland (1.1 GB)

#> Downloading http://download.geofabrik.de/europe/poland-latest.osm.pbf to 
#> /tmp/RtmpWjPinl/Poland.osm.pbf

#> Old attributes: attributes=name,highway,waterway,aerialway,barrier,man_made

#> New attributes: attributes=name,highway,waterway,aerialway,barrier,man_made,maxspeed,oneway,building,surface,landuse,natural,start_date,wall,service,lanes,layer,tracktype,bridge,foot,bicycle,lit,railway,footway

#> Using ini file that can can be edited with file.edit(/tmp/RtmpWjPinl/ini_new.ini)

#> Warning: GDAL Error 1: Too many features have accumulated in points layer. Use OGR_INTERLEAVED_READING=YES mode

# get zone associated with a point
name = sf::st_sfc(sf::st_point(c(0, 53)), crs = 4326)
get_geofabric(name)

#> Error in if (nrow(name) > 1) {    warning("Matching only based on the first feature.")    message("Try sf::st_union() to convert into a single multi feature.")}: argument is of length zero

name = sf::st_sfc(sf::st_point(c(0, 53)), sf::st_point(c(-2, 55)), crs = 4326)
gf_find_sf(name)

#> although coordinates are longitude/latitude, st_contains assumes that they are planar

#> The place is within these geofabrik zones: Europe, Great Britain, Britain and Ireland, England, Lincolnshire, Northumberland
#> Selecting the smallest: Lincolnshire

#> Simple feature collection with 1 feature and 11 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XYZ
#> bbox:           xmin: -0.950922 ymin: 52.63954 xmax: 0.494382 ymax: 53.73003
#> epsg (SRID):    4326
#> proj4string:    +proj=longlat +datum=WGS84 +no_defs
#>             name  size_pbf
#> 401 Lincolnshire (25.6 MB)
#>                                                                                   pbf_url
#> 401 http://download.geofabrik.de/europe/great-britain/england/lincolnshire-latest.osm.pbf
#>                                                                        page_url
#> 401 http://download.geofabrik.de/europe/great-britain/england/lincolnshire.html
#>     part_of level continent country  region    subregion
#> 401 England     4      <NA>    <NA> England Lincolnshire
#>                                                                   geometry_url
#> 401 http://download.geofabrik.de/europe/great-britain/england/lincolnshire.kml
#>                           geometry
#> 401 MULTIPOLYGON Z (((-0.761741...

# }